tools/requirements.txt
google-cloud-translate==2.0.1